veritasai/requirements.txt
streamlit
requests
feedparser
newspaper3k
lxml
lxml_html_clean
transformers
faster-whisper
numpy
tqdm
bcrypt
python-dotenv
openai
scikit-learn
pandas
matplotlib
seaborn
plotly
Pillow
scipy
nltk
librosa 